small cleanup before sharing with JB
