+++ consumer/src/main/java/arrow/sample/test.kt
@@ -5,7 +5,7 @@ package arrow.sample
 object test {
   @JvmStatic
   fun main(args: Array<String>) {
-    println(Option::class.java.interfaces.toList().map { it.toGenericString() })
-    println(Class.forName("arrow.sample.ForOption"))
+    println("Option supertypes: ${Option::class.java.interfaces.toList().map { it.toGenericString() }}")
+    println("For Option Class through reflection:" + Class.forName("arrow.sample.ForOption"))
   }
 }
